Fix bug where clicking results after searching would require 2 clicks
After typing in the search input, the "change" event listener was swapping the results. The mousedown on the result would blur the input and trigger the change event thus swapping the results. The click on the result would fail to complete because the mouseup event was not on the same element as the mousedown.