Fix bug where clearing a prepopulating search.
Because prepopulating a search just sets the value of the search input with actually updating the query, the query is still empty, so clearing it doesn't actually modify it. This means the query change event doesn't fire, meaning the updateClearSearchButtonVisiblity function doesn't run. So run it manually when the clear search button is clicked.