Add an onRender callback to allow manipulation of rendered results.
This makes it possible to manipulate results before the selected one is scrolled into view. As with onSelect, passing in a callback instead of attaching an event handler seems to make more sense because it is more natural for a callback to reference internal view elements than it is for an event listener to.