Set correct value when refreshing select options
Browsers do not update the <select> element's value when its <option> elements are changed. This means if we ask the select what its value is, it will not necessarily give the value of the selected option element, as it returns whatever the previously set value was.

Instead, only when refreshing select options, we find the selected <option> element and use its value to update the <select> element value and the UberSearch value.