Fix bug where disabled attribute was not detected
If the <option> element only had a `disabled` attribute without a value it would be interpreted as a falsey value since we're not using jQuery to read attributes for performance reasons.

Instead, we look for a value of `undefined` and treat it as `false`. Anything else will be considered `true`. This works because if the attribute is not present we ensure that `getAttribute` returns `undefined`.