Fix bug where an unwanted change event could trigger
We were comparing the given value to the value from the select, however, some values are coerced to a normalized form by the `select.val()` method, so something like `undefined` would become emptystring. This means that we would always fire a change event if we successively assigned `undefined`, `null`, or other values that are coerced.

Instead, we read the before and after value using `select.val()` so we always compare coerced values.