Allow arrow keys focus from trigger to results
We lost this ability in https://github.com/culturecode/uber_select/commit/b4b173d6ebee841acb0337b818e7ff51441d9407 when we stopped listening to the trigger after opening the result pane.

Instead, we now focus the first result when we open the pane. We also listen for the down key on the trigger when the pane is open so if the user shift-tabs back to the trigger, they can press down to go back to the results.