Simplify keyboard event handling
Instead of changing where we listen for keyboard events depending on whether there is a search input or not, we now listen to the container that contains both the search input and the `List`. This relies on keyboard events like up/down arrow presses being cancelled if they are handled by the `List`.