Fix bug when opening pane
In 693566493e1dc7b133e94d42f42e92a56c545ccd I added a namespace to the “shown” and “hidden” `Pane` events to avoid such common events from polluting the events. This was a premature move and caused a bug where we were not listening for the namespaced event.

Instead, I’m reverting the event name change to fix the bug it introduced.