Add support for matching group names
`options.matchGroupNames` can be used to control whether the search matches against group names. A matched group name will cause results within that group to be returned. This is acheived by modifying how the `datumPreprocessor` function works, so it a custom `datumPreprocessor` is passed in, this option will have no effect. Since `datumPreprocessor` is not described in the readme, I’m not going to worry about this conflict at this time.