Focus on list when down arrow pressed in search input
When the down arrow is pressed inside the search input, then we want to focus on the list of results.

Remove events that were not being triggered. The pane is never in focus, so these events were never being triggered

Reimplement correct focusing when pane is open. When a user uses the up arrow to get back to the outerContainer, if they use the down arrow, they should be able to select the options.