Update aria roles for uberselect
Move the listbox role down to the pane. Per the examples in the documentation, the select label is not included in the listbox.

Add listitem role to each list item so screen-readers know the options that are inside the select.

Reference for example of listbox html: https://www.w3.org/WAI/ARIA/apg/patterns/listbox/examples/listbox-grouped/