Support use in Shadow DOM
If the Uber Select is used as part of an element in the shadow DOM, all events will appear to originate from the host element. This means they will all look like they came from outside the Uber Select and it will immediately close.

Instead, we detect a composed event and search the composed path for the view.